+++ services/sts/sts-core/src/main/java/org/apache/cxf/sts/token/provider/SymmetricKeyHandler.java
@@ -109,8 +109,8 @@ public class SymmetricKeyHandler {
             } else if (STSConstants.SYMMETRIC_KEY_TYPE.equals(binarySecret.getBinarySecretType())
                 || binarySecret.getBinarySecretType() == null) {
                 byte[] secretValue = binarySecret.getBinarySecretValue();
-                if (((long)secretValue.length * 8L) < signatureProperties.getMinimumKeySize()
-                    || ((long)secretValue.length * 8L) > signatureProperties.getMaximumKeySize()) {
+                if ((secretValue.length * 8L) < signatureProperties.getMinimumKeySize()
+                    || (secretValue.length * 8L) > signatureProperties.getMaximumKeySize()) {
                     LOG.log(
                         Level.WARNING, "Received secret of length " + secretValue.length
                         + " bits is not accepted"
@@ -128,8 +128,8 @@ public class SymmetricKeyHandler {
             }
         } else if (clientEntropy.getDecryptedKey() != null) {
             byte[] secretValue = clientEntropy.getDecryptedKey();
-            if (((long)secretValue.length * 8L) < signatureProperties.getMinimumKeySize()
-                || ((long)secretValue.length * 8L) > signatureProperties.getMaximumKeySize()) {
+            if ((secretValue.length * 8L) < signatureProperties.getMinimumKeySize()
+                || (secretValue.length * 8L) > signatureProperties.getMaximumKeySize()) {
                 LOG.log(
                     Level.WARNING, "Received secret of length " + secretValue.length
                     + " bits is not accepted"
